StationApplication/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
Deprecated==1.2.13
Django==4.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-oauth-toolkit==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.7.1
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
